Do not re-export _Unwind_XXX symbols.
The problem is that this library is getting _Unwind_XXX symbols from
libgcc.a for the arm targets, and then re-exporting them.

This is causing facebook to crash because a shared library is using
an undefined reference to an unwind symbol and then using the one
from libz.so, which is mixing an llvm unwinder with a gcc one.

Test: facebook no longer crashes when starting up.
Change-Id: I1c2439f08375fbc630a49fdabcc50d22d461eb2b
(cherry picked from commit 4e652aee9d360aac8cfc2812b644a25d6ba7840f)