Re-enable DownloadProtectionServiceBrowserTest.VerifyZipHash
This CL removes the check in ZipReader that the target file is
empty. We can't truncate the file within the sandbox on MacOS,
so we're forced to overwrite the first few bytes for each
entry in the zip. This was recently forbidden by ZipReader, so
reverse that part of the change.

This CL also adds a few lines that would have made debugging this
easier. In the future, instead of just knowing the hash of b.zip is
e3b0c4..., we'll be explicitly told the entry was empty and
truncation failed.

Fixed: 1309879
Change-Id: Idd34436df21f55f643b12d024401b7cb7e3e4138
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3563962
Reviewed-by: Rohit Bhatia <bhatiarohit@google.com>
Commit-Queue: Daniel Rubery <drubery@chromium.org>
Reviewed-by: François Degros <fdegros@chromium.org>
Cr-Commit-Position: refs/heads/main@{#987758}
NOKEYCHECK=True
GitOrigin-RevId: 458293487a3211c5f15882d3fe3e4fecbddad333