[zlib] Enable LIT_MEM code path
Introduced on zlib 1.3.0.1, this will allow a small compression
speed improvement at cost of using a bit more memory.

Observed average gain was 2.6% (SPR) and 2.3% (M1).

It is interesting that it will have a bigger impact on files with higher entropy like JPG (+5.3%) and pre-compressed (baddata +7%), probably because those are cases where more literals are produced by the encoder since it is harder to find matches in LZ77 for such inputs.

Bug: 1447443
Change-Id: I09f15c8bd0f400d5485a10aa38bdc2dd1ece2449
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5078349
Reviewed-by: Hans Wennborg <hans@chromium.org>
Commit-Queue: Adenilson Cavalcanti <cavalcantii@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1237242}
NOKEYCHECK=True
GitOrigin-RevId: 0f445392841590a53180bf4c4e9a9d9b9d627c6a