Allow //bootable/recovery/updater to use zlib_stable.
Some updater tests fail if run against the latest version of zlib
because of a CRC change. zlib_stable (as its name implies) opts out
of any such "fancy new stuff", but we need updater to opt _in_ to
using that version of zlib. The other OTA tools already do.

Test: treehugger
Change-Id: I3707b940f724f62837a0b57e22ff315b36306e37